streamlit/requirements.txt
streamlit==1.44.1
opencv-python==4.11.0.86
torch==2.6.0
torchvision==2.6.0
ultralytics==8.3.118
numpy==2.0.2
Pillow==11.1.0
streamlit-webrtc==0.57.0
playsound3
fpdf